renames core package
