+++ apps/angular-example-counter/src/app/store/effects/index.ts
@@ -1,5 +1,5 @@
 import { scan, startWith } from 'rxjs/operators';
-import { Effect } from '@rx-store/rx-store';
+import { Effect } from '@rx-store/core';
 import { RootAppStore } from '../../types';
 
 /**
@@ -19,11 +19,10 @@ import { RootAppStore } from '../../types';
  * by accessing the context value directly in your components, and
  * subscribing. See the <Provider /> component for an example.
  */
-export const appRootEffect: Effect<RootAppStore> = ({sources, sinks}) => {
-  return sources.counterChange$()
-    .pipe(
-      scan((acc, val) => acc + val, 0),
-      startWith(0),
-      sinks.count$()
-    )
+export const appRootEffect: Effect<RootAppStore> = ({ sources, sinks }) => {
+  return sources.counterChange$().pipe(
+    scan((acc, val) => acc + val, 0),
+    startWith(0),
+    sinks.count$()
+  );
 };
