+++ apps/angular-example-counter/src/app/store/store.service.ts
@@ -1,11 +1,10 @@
 import { Injectable, OnDestroy } from '@angular/core';
-import { Effect } from '@rx-store/rx-store';
+import { Effect } from '@rx-store/core';
 
 @Injectable({
-  providedIn: 'root'
+  providedIn: 'root',
 })
 export class StoreService implements OnDestroy {
-
   appEffect: any;
 
   init(storeValue: any, storeEffect: Effect<any>): void {
@@ -13,9 +12,8 @@ export class StoreService implements OnDestroy {
   }
 
   ngOnDestroy(): void {
-    if (this.appEffect){
+    if (this.appEffect) {
       this.appEffect();
     }
   }
-
 }
