+++ apps/react-example-counter/src/app/app-context-value.interface.ts
@@ -1,5 +1,5 @@
 import { Subject, BehaviorSubject } from 'rxjs';
-import { StoreValue } from '@rx-store/rx-store';
+import { StoreValue } from '@rx-store/core';
 
 /** Our context value, which contains subjects & observables */
 export interface AppContextValue extends StoreValue {
