+++ apps/rx-store-website/docs/introduction/installation.md
@@ -1,6 +1,6 @@
 ---
 id: installation
-title: "Installation"
+title: 'Installation'
 ---
 
 You will likely want to use the React specific installation instructions, but if you're
@@ -9,11 +9,11 @@ using this outside of React, you can use the standalone core code / types:
 ## NPM
 
 ```
-npm i @rx-store/rx-store rxjs
+npm i @rx-store/core rxjs
 ```
 
 ## Yarn
 
 ```
-yarn add @rx-store/rx-store rxjs
+yarn add @rx-store/core rxjs
 ```
