+++ libs/core/.babelrc
