+++ libs/core/.eslintrc
