+++ libs/core/README.md
