+++ libs/core/jest.config.js
