+++ libs/core/src/lib/store-value.ts
