+++ libs/core/tsconfig.json
