+++ libs/core/tsconfig.lib.json
