+++ libs/core/tsconfig.spec.json
