+++ libs/react-rx-store/package.json
@@ -3,7 +3,7 @@
   "version": "0.0.4",
   "license": "ISC",
   "dependencies": {
-    "@rx-store/rx-store": "^0.0.4",
+    "@rx-store/core": "^0.0.4",
     "debug": "^4.1.1"
   }
 }
