+++ tsconfig.json
@@ -16,7 +16,7 @@
     "skipDefaultLibCheck": true,
     "baseUrl": ".",
     "paths": {
-      "@rx-store/rx-store": ["libs/rx-store/src/index.ts"],
+      "@rx-store/core": ["libs/core/src/index.ts"],
       "@rx-store/react-rx-store": ["libs/react-rx-store/src/index.ts"]
     }
   },
